함수간 띄워쓰기 적용 in -> on으로 변경

+++ UnlimitedTowerContract/Contract_eosiocpp/Table/battle_table.hpp
@@ -11,8 +11,8 @@ enum ebattle_member_state
 // 1 + 8 = 9
 struct sbattle_member_state
 {
-    uint8_t turn_count;       //캐릭터 상태의 지속 턴 횟수
-    uint64_t turn_state;           //캐릭터의 현재 상태
+    uint32_t turn_count;       //캐릭터 상태의 지속 턴 횟수
+    uint32_t turn_state;           //캐릭터의 현재 상태
 };
 // 4 + 4 + 4 + 4 + 4 + 4 + 8 + sbattle_member_state(9) = 41
 // sbattle_member_state 당 9 총 5개의 버프창이 있으면 45 + 32 = 77
